feat: Update character and NPC data structures to use int for directional coordinates

+++ Game.DTOs/Game/Npc/NpcData.cs
@@ -5,7 +5,7 @@ namespace Game.DTOs.Game.Npc;
 [MemoryPackable]
 public readonly partial record struct NpcData(
     int NpcId, int NetworkId, int MapId, string Name,
-    int X, int Y, int Z, sbyte DirX, sbyte DirY,
+    int X, int Y, int Z, int DirX, int DirY,
     int Hp, int MaxHp, float HpRegen,
     int Mp, int MaxMp, float MpRegen,
     float MovementSpeed, float AttackSpeed,
