+++ Game.DTOs/Game/Player/PlayerData.cs
@@ -7,7 +7,7 @@ public readonly partial record struct PlayerData(
     int PlayerId, int NetworkId, int MapId,
     string Name, byte Gender, byte Vocation,
     int X, int Y, int Z,
-    sbyte DirX, sbyte DirY,
+    int DirX, int DirY,
     int Hp, int MaxHp, float HpRegen,
     int Mp, int MaxMp, float MpRegen,
     float MovementSpeed, float AttackSpeed,
