+++ Game.Domain/Entities/Character.cs
@@ -1,4 +1,3 @@
-// Domain/Entities/Character.cs
 using Game.Domain.Enums;
 
 namespace Game.Domain.Entities;
@@ -15,8 +14,8 @@ public sealed class Character : BaseEntity
     public VocationType Vocation { get; set; } = VocationType.Unknown;
     
     // Posição no mundo
-    public sbyte FacingX { get; set; }
-    public sbyte FacingY { get; set; }
+    public int FacingX { get; set; }
+    public int FacingY { get; set; }
     public int PositionX { get; set; }
     public int PositionY { get; set; }
     public int PositionZ { get; set; }
